+++ sdk/src/main/kotlin/com/processout/sdk/api/model/response/napm/v2/PONativeAlternativePaymentAuthorizationResponse.kt
@@ -1,8 +1,5 @@
 package com.processout.sdk.api.model.response.napm.v2
 
-import com.processout.sdk.api.model.response.POBarcode
-import com.processout.sdk.api.model.response.POImageResource
-import com.processout.sdk.api.model.response.napm.v2.PONativeAlternativePaymentAuthorizationResponse.CustomerInstruction
 import com.processout.sdk.api.model.response.napm.v2.PONativeAlternativePaymentAuthorizationResponse.State
 import com.processout.sdk.core.annotation.ProcessOutInternalApi
 import com.squareup.moshi.Json
@@ -20,7 +17,7 @@ import com.squareup.moshi.JsonClass
 data class PONativeAlternativePaymentAuthorizationResponse(
     val state: State,
     val nextStep: PONativeAlternativePaymentNextStep?,
-    val customerInstructions: List<CustomerInstruction>?
+    val customerInstructions: List<PONativeAlternativePaymentCustomerInstruction>?
 ) {
 
     /**
@@ -44,75 +41,6 @@ data class PONativeAlternativePaymentAuthorizationResponse(
         @ProcessOutInternalApi
         UNKNOWN
     }
-
-    /**
-     * Specifies instruction for the customer, providing additional information and/or describing required actions.
-     */
-    sealed class CustomerInstruction {
-
-        /**
-         * Customer instruction provided as a markdown text.
-         *
-         * @param[label] Optional instruction display label.
-         * @param[value] Markdown text.
-         */
-        @JsonClass(generateAdapter = true)
-        data class Text(
-            val label: String?,
-            val value: String
-        ) : CustomerInstruction()
-
-        /**
-         * Customer instruction provided as an image resource.
-         *
-         * @param[value] Image resource.
-         */
-        @JsonClass(generateAdapter = true)
-        data class Image(
-            val value: POImageResource
-        ) : CustomerInstruction()
-
-        /**
-         * Customer instruction provided via barcode.
-         *
-         * @param[rawSubtype] Raw barcode subtype.
-         * @param[rawValue] Base64 encoded value.
-         */
-        @JsonClass(generateAdapter = true)
-        data class Barcode(
-            @Json(name = "subtype")
-            val rawSubtype: String,
-            @Json(name = "value")
-            val rawValue: String
-        ) : CustomerInstruction() {
-
-            /** Barcode value. */
-            val value: POBarcode
-                get() = POBarcode(
-                    rawType = rawSubtype,
-                    rawValue = rawValue
-                )
-        }
-
-        /**
-         * Group of customer instructions.
-         *
-         * @param[label] Optional group display label.
-         * @param[instructions] Grouped instructions for the customer that provide additional information and/or describe required actions.
-         */
-        @JsonClass(generateAdapter = true)
-        data class Group(
-            val label: String?,
-            val instructions: List<CustomerInstruction>
-        ) : CustomerInstruction()
-
-        /**
-         * Placeholder that allows adding additional cases while staying backward compatible.
-         * __Warning:__ Do not match this case directly, use _when-else_ instead.
-         */
-        @ProcessOutInternalApi
-        data object Unknown : CustomerInstruction()
-    }
 }
 
 @JsonClass(generateAdapter = true)
@@ -121,5 +49,5 @@ internal data class NativeAlternativePaymentAuthorizationResponseBody(
     @Json(name = "next_step")
     val nextStep: NativeAlternativePaymentNextStep?,
     @Json(name = "customer_instructions")
-    val customerInstructions: List<CustomerInstruction>?
+    val customerInstructions: List<PONativeAlternativePaymentCustomerInstruction>?
 )
